Modified discrete_act.py in order to support n_bins > 3 (#40)
* Modified discrete_act.py in order to support n_bins > 3

* Faster method of casting

+++ rlgym/utils/action_parsers/discrete_act.py
@@ -18,9 +18,9 @@ class DiscreteAction(ActionParser):
         return gym.spaces.MultiDiscrete([self._n_bins] * 5 + [2] * 3)
 
     def parse_actions(self, actions: np.ndarray, state: GameState) -> np.ndarray:
-        actions = actions.reshape((-1, 8))
+        actions = actions.reshape((-1, 8)).astype(dtype=np.float32)
 
-        # map all ternary actions from {0, 1, 2} to {-1, 0, 1}.
+        # map all binned actions from {0, 1, 2 .. n_bins - 1} to {-1 .. 1}.
         actions[..., :5] = actions[..., :5] / (self._n_bins // 2) - 1
 
         return actions
